+++ projects/mk-kit/src/lib/components/directives/index.ts
@@ -3,3 +3,5 @@ export * from './copy-to-clipboard';
 export * from './autosize';
 export * from './autofocus';
 export * from './scrollspy';
+export * from './intersect';
+export * from './infinite-scroll';